FIX Decimal 'D' format scale 0 being omitted in binary
Decimal values being represented in the 'D' format that have a scale of 0
were being omitted in the binary slice representation. This matches the
behaviour in the other Ion libraries. The issue was being caused because
with the omission of the scale 0, the dot binary representation in the slice was
being mistaken as the scale value.