docs: add users section to readme
This small change adds a "users" section to the README that can point at open source projects written in Go that support Amazon Ion by using this library. I've added one of my projects as an example with some text to encourage others to do the same.

Why link to these external projects?

1. They provide a real-world example of using the library that others can learn from
2. They provide tools that may be useful in testing or adopting Amazon Ion
3. They can help create community and a useful feedback mechanism for bugs & features

A little background on my tool: it's a CLI for HTTP APIs that has built-in content negotiation features and supports a bunch of content types out of the box, so for example if the API server sends binary Amazon Ion instead of plain JSON then Restish can parse and display the resulting data.