Correct the coefficient value of timestamp nanoseconds
Writing a timestamp nanosecond into Ion binary will write bytes of two values: Coefficient and exponent. These two values are used to calculate the nanoseconds using the equation: coefficient * 10 ^ -exponent.

For example a timestamp:  `2021-05-27 00:19:05.000050000` has coefficient: `50000` and exponent: `9`.

Inside marshal binary function, we get the correct coefficient by calling `time.Nanosecond` which returns `50000` however the exponent is calculated by converting the coefficient to string and getting the length which returns `5` and is incorrect.

We can set the exponent to `9` if there are nanoseconds because it will always be `9` precision units.

The next step is to read the binary ion timestamps correctly. Original implementation was to look at the cofficient value and remove trailing zeros but the binary ion already contains a byte containing the exponent so we can just use that.