Set fractional precision to 9 if timestamp precision is TimestampPrecisionNanosecond
When we create a timestamp object from Golang's time object, we were specifying the precision = 9 if nanoseconds > 0. This is incorrect because according to Go time docs, time object represents an instant in time with nanosecond precision.

The change is to set fractional precision = 9 when creating a timestamp object from Go time object.

Another change added was to add a guard to timestamp creation functions and set fractional precision = 9 if timestamp precision =  TimestampPrecisionNanosecond. Otherwise set it to 0.